# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Legacy Sports API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pricing:
    get:
      operationId: getPricing
      tags: [billing]
      summary: Get platform pricing configuration (superadmin)
      responses:
        "200":
          description: Pricing config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingConfig"
    put:
      operationId: updatePricing
      tags: [billing]
      summary: Update platform pricing configuration (superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingConfigInput"
      responses:
        "200":
          description: Updated pricing config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingConfig"

  # ── SUBSCRIPTIONS ──────────────────────────────────────────────────────────
...
  /subscriptions/{schoolId}:
    patch:
      operationId: updateSubscription
      tags: [billing]
      summary: Update subscription settings for a school (superadmin)
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{schoolId}/preview:
    get:
      operationId: previewInvoice
      tags: [billing]
      summary: Preview the next invoice for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoicePreview"
  /subscriptions/{schoolId}/invoice:
    post:
      operationId: generateInvoice
      tags: [billing]
      summary: Generate an invoice for a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ── INVOICES ───────────────────────────────────────────────────────────────
  /invoices:
    get:
      operationId: listInvoices
      tags: [billing]
      summary: List invoices
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [billing]
      summary: Override invoice total, due date, or notes (superadmin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/pay:
    post:
      operationId: markInvoicePaid
      tags: [billing]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/unpay:
    post:
      operationId: markInvoiceUnpaid
      tags: [billing]
      summary: Reverse a payment (bounced/refunded)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/void:
    post:
      operationId: voidInvoice
      tags: [billing]
      summary: Void an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voided invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/remind:
    post:
      operationId: sendInvoiceReminders
      tags: [billing]
      summary: Send AI payment reminders for an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                channels:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: Reminder logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderLog"

  # ── REMINDERS ──────────────────────────────────────────────────────────────
  /reminders:
    get:
      operationId: listReminders
      tags: [billing]
      summary: List reminder logs
      parameters:
        - name: invoiceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reminder logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReminderLog"

  # ── BILLING CYCLE ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PricingConfig:
      type: object
      required:
        - id
        - baseFee
        - perPlayerFee
        - perSportFee
        - modAttendance
        - modPerformance
        - modAnalytics
        - modLeaderboard
        - modCalendar
        - modMessaging
        - modPhotos
        - modNotifications
        - modFees
        - modRegistration
        - modAi
        - modWebsite
        - currency
        - gracePeriodDays
        - reminderDaysBefore
        - autoSuspendAfterDays
      properties:
        id:
          type: integer
        baseFee:
          type: integer
          description: Base subscription fee per month
        perPlayerFee:
          type: integer
          description: Fee per active player per month
        perSportFee:
          type: integer
          description: Fee per enabled sport per month
        modAttendance:
          type: integer
        modPerformance:
          type: integer
        modAnalytics:
          type: integer
        modLeaderboard:
          type: integer
        modCalendar:
          type: integer
        modMessaging:
          type: integer
        modPhotos:
          type: integer
        modNotifications:
          type: integer
        modFees:
          type: integer
        modRegistration:
          type: integer
        modAi:
          type: integer
        modWebsite:
          type: integer
        currency:
          type: string
        gracePeriodDays:
          type: integer
        reminderDaysBefore:
          type: integer
        autoSuspendAfterDays:
          type: integer
        upiId:
          type: string
        qrCodeUrl:
          type: string
        account1BankName:
          type: string
        account1AccountNumber:
          type: string
        account1IfscCode:
          type: string
        account1HolderName:
          type: string
        account2BankName:
          type: string
        account2AccountNumber:
          type: string
        account2IfscCode:
          type: string
        account2HolderName:
          type: string
        updatedAt:
          type: string
          format: date-time
    PricingConfigInput:
      type: object
      properties:
        baseFee:
          type: integer
        perPlayerFee:
          type: integer
        perSportFee:
          type: integer
        modAttendance:
          type: integer
        modPerformance:
          type: integer
        modAnalytics:
          type: integer
        modLeaderboard:
          type: integer
        modCalendar:
          type: integer
        modMessaging:
          type: integer
        modPhotos:
          type: integer
        modNotifications:
          type: integer
        modFees:
          type: integer
        modRegistration:
          type: integer
        modAi:
          type: integer
        modWebsite:
          type: integer
        currency:
          type: string
        gracePeriodDays:
          type: integer
        reminderDaysBefore:
          type: integer
        autoSuspendAfterDays:
          type: integer
        upiId:
          type: string
        qrCodeUrl:
          type: string
        account1BankName:
          type: string
        account1AccountNumber:
          type: string
        account1IfscCode:
          type: string
        account1HolderName:
          type: string
        account2BankName:
          type: string
        account2AccountNumber:
          type: string
        account2IfscCode:
          type: string
        account2HolderName:
          type: string
    Subscription:
      type: object
      required: [id, schoolId, planName, status, autoBill]
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        planName:
          type: string
        billingCycleDay:
          type: integer
        status:
          type: string
        autoBill:
          type: boolean
        nextInvoiceDate:
          type: string
          format: date
        lastInvoicedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubscriptionInput:
      type: object
      properties:
        planName:
          type: string
        billingCycleDay:
          type: integer
        status:
          type: string
        autoBill:
          type: boolean
        nextInvoiceDate:
          type: string
          format: date
    InvoiceLineItem:
      type: object
      required: [label, qty, unit, total]
      properties:
        label:
          type: string
        qty:
          type: number
        unit:
          type: number
        total:
          type: number
    Invoice:
      type: object
      required: [id, schoolId, invoiceNumber, periodStart, periodEnd, dueDate, subtotal, total, currency, status]
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        invoiceNumber:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        subtotal:
          type: number
        total:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, overdue, void]
        paidAt:
          type: string
          format: date-time
        paidMethod:
          type: string
        notes:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        createdAt:
          type: string
          format: date-time
    InvoiceUpdate:
      type: object
      properties:
        total:
          type: number
        subtotal:
          type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string
        status:
          type: string
    InvoicePreview:
      type: object
      required: [subtotal, total, currency, playerCount]
      properties:
        subtotal:
          type: number
        total:
          type: number
        currency:
          type: string
        playerCount:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
    ReminderLog:
      type: object
      required: [id, invoiceId, channel, status, createdAt]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        schoolId:
          type: integer
        channel:
          type: string
        recipient:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        provider:
          type: string
        error:
          type: string
        createdAt:
          type: string
          format: date-time
...